@@ -1,8 +1,15 @@
 module github.com/suzaku/shonenjump
 
-go 1.12
+go 1.19
 
 require (
 	github.com/mattn/go-isatty v0.0.12
 	github.com/stretchr/testify v1.6.1
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	golang.org/x/sys v0.1.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
 )
